backend/requirements.txt

fastapi
uvicorn[standard]
python-multipart
websockets
psycopg2-binary
sqlalchemy
alembic
supabase
httpx
requests
bcrypt
python-jose[cryptography]
passlib[bcrypt]
cryptography
python-dotenv
pydantic
pydantic-settings
openai
anthropic
duckduckgo-search
yfinance
pandas
numpy
loguru
structlog
redis
celery
pytest
pytest-asyncio
black
isort
flake8
gunicorn
asyncpg
agno
psutil
prometheus_client
agno[anthropic,openai,groq,yfinance,ddg,reasoning]
aiofiles
slowapi
chromadb
PyPDF2
python-docx
python-magic
tenacity
numpy